Move tests to separate files; add track_abundance fixture for checking both implementations
