+++ sourmash_lib/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture(params=[True, False])
+def track_abundance(request):
+    return request.param
