+++ sourmash_lib/signature.py
@@ -256,80 +256,3 @@ def save_signatures(siglist, fp=None):
             s += '---\n'
 
     return s
-
-
-def test_roundtrip():
-    e = sourmash_lib.Estimators(n=1, ksize=20)
-    e.add("AT" * 10)
-    sig = SourmashSignature('titus@idyll.org', e)
-    s = save_signatures([sig])
-    siglist = list(load_signatures(s))
-    sig2 = siglist[0]
-    e2 = sig2.estimator
-
-    assert sig.similarity(sig2) == 1.0
-    assert sig2.similarity(sig) == 1.0
-
-
-def test_roundtrip_empty_email():
-    e = sourmash_lib.Estimators(n=1, ksize=20)
-    e.add("AT" * 10)
-    sig = SourmashSignature('', e)
-    s = save_signatures([sig])
-    siglist = list(load_signatures(s))
-    sig2 = siglist[0]
-    e2 = sig2.estimator
-
-    assert sig.similarity(sig2) == 1.0
-    assert sig2.similarity(sig) == 1.0
-
-
-def test_md5():
-    e = sourmash_lib.Estimators(n=1, ksize=20)
-    e.mh.add_hash(5)
-    sig = SourmashSignature('titus@idyll.org', e)
-    print(sig._save())
-    assert sig.md5sum() == 'eae27d77ca20db309e056e3d2dcd7d69', sig.md5sum()
-
-
-def test_name():
-    e = sourmash_lib.Estimators(n=1, ksize=20)
-    sig = SourmashSignature('titus@idyll.org', e, name='foo')
-    assert sig.name() == 'foo'
-
-
-def test_name_2():
-    e = sourmash_lib.Estimators(n=1, ksize=20)
-    sig = SourmashSignature('titus@idyll.org', e, filename='foo.txt')
-    assert sig.name() == 'foo.txt'
-
-
-def test_name_3():
-    e = sourmash_lib.Estimators(n=1, ksize=20)
-    sig = SourmashSignature('titus@idyll.org', e, name='foo',
-                            filename='foo.txt')
-    assert sig.name() == 'foo'
-
-
-def test_name_4():
-    e = sourmash_lib.Estimators(n=1, ksize=20)
-    sig = SourmashSignature('titus@idyll.org', e)
-    assert sig.name() == sig.md5sum()[:8]
-
-
-def test_save_load_multisig():
-    e1 = sourmash_lib.Estimators(n=1, ksize=20)
-    sig1 = SourmashSignature('titus@idyll.org', e1)
-    
-    e2 = sourmash_lib.Estimators(n=1, ksize=20)
-    sig2 = SourmashSignature('titus2@idyll.org', e2)
-
-    x = save_signatures([sig1, sig2])
-    y = list(load_signatures(x))
-
-    print(x)
-
-    assert len(y) == 2
-    assert sig1 in y                      # order not guaranteed, note.
-    assert sig2 in y
-    assert sig1 != sig2
