+++ sourmash_lib/test__minhash.py
@@ -37,7 +37,6 @@ from __future__ import print_function
 from __future__ import absolute_import, unicode_literals
 
 from ._minhash import MinHash, hash_murmur
-import screed
 
 # add:
 # * get default params from Python
@@ -47,9 +46,9 @@ import screed
 # * nan on empty minhash
 # * define equals
 
-def test_basic_dna():
+def test_basic_dna(track_abundance):
     # verify that MHs of size 1 stay size 1, & act properly as bottom sketches.
-    mh = MinHash(1, 4)
+    mh = MinHash(1, 4, track_abundance=track_abundance)
     mh.add_sequence('ATGC')
     a = mh.get_mins()
 
@@ -61,25 +60,25 @@ def test_basic_dna():
     assert len(b) == 1
 
 
-def test_protein():
+def test_protein(track_abundance):
     # verify that we can hash protein/aa sequences
-    mh = MinHash(10, 6, True)
+    mh = MinHash(10, 6, True, track_abundance=track_abundance)
     mh.add_protein('AGYYG')
 
     assert len(mh.get_mins()) == 4
 
 
-def test_protein_short():
+def test_protein_short(track_abundance):
     # verify that we can hash protein/aa sequences
-    mh = MinHash(10, 9, True)
+    mh = MinHash(10, 9, True, track_abundance=track_abundance)
     mh.add_protein('AG')
 
     assert len(mh.get_mins()) == 0, mh.get_mins()
 
 
-def test_basic_dna_bad():
+def test_basic_dna_bad(track_abundance):
     # test behavior on bad DNA
-    mh = MinHash(1, 4)
+    mh = MinHash(1, 4, track_abundance=track_abundance)
     try:
         mh.add_sequence('ATGR')
         assert 0, "should fail on invalid DNA sequence"
@@ -87,9 +86,9 @@ def test_basic_dna_bad():
         pass
 
 
-def test_basic_dna_bad_2():
+def test_basic_dna_bad_2(track_abundance):
     # test behavior on bad DNA
-    mh = MinHash(1, 6)
+    mh = MinHash(1, 6, track_abundance=track_abundance)
     try:
         mh.add_protein('YYYY')
         assert 0, "should fail => this is a DNA MinHash"
@@ -97,9 +96,9 @@ def test_basic_dna_bad_2():
         pass
 
 
-def test_basic_dna_bad_force():
+def test_basic_dna_bad_force(track_abundance):
     # test behavior on bad DNA
-    mh = MinHash(1, 4)
+    mh = MinHash(1, 4, track_abundance=track_abundance)
     assert len(mh.get_mins()) == 0
     mh.add_sequence('ATGR', True)
     assert len(mh.get_mins()) == 1
@@ -109,9 +108,9 @@ def test_basic_dna_bad_force():
     assert len(mh.get_mins()) == 1
 
 
-def test_compare_1():
-    a = MinHash(20, 10)
-    b = MinHash(20, 10)
+def test_compare_1(track_abundance):
+    a = MinHash(20, 10, track_abundance=track_abundance)
+    b = MinHash(20, 10, track_abundance=track_abundance)
 
     a.add_sequence('TGCCGCCCAGCACCGGGTGACTAGGTTGAGCCATGATTAACCTGCAATGA')
     b.add_sequence('TGCCGCCCAGCACCGGGTGACTAGGTTGAGCCATGATTAACCTGCAATGA')
@@ -128,53 +127,53 @@ def test_compare_1():
     assert b.compare(a) == 1.0
     assert a.compare(a) == 1.0
 
-    
+
     b.add_sequence('GATTGGTGCACACTTAACTGGGTGCCGCGCTGGTGCTGATCCATGAAGTT')
     x = a.compare(b)
     assert x >= 0.3, x
-    
+
     x = b.compare(a)
     assert x >= 0.3, x
     assert a.compare(a) == 1.0
     assert b.compare(b) == 1.0
 
 
-def test_mh_copy():
-    a = MinHash(20, 10)
+def test_mh_copy(track_abundance):
+    a = MinHash(20, 10, track_abundance=track_abundance)
 
     a.add_sequence('TGCCGCCCAGCACCGGGTGACTAGGTTGAGCCATGATTAACCTGCAATGA')
     b = a.__copy__()
     assert b.compare(a) == 1.0
 
 
-def test_mh_len():
-    a = MinHash(20, 10)
+def test_mh_len(track_abundance):
+    a = MinHash(20, 10, track_abundance=track_abundance)
 
     assert len(a) == 20
     a.add_sequence('TGCCGCCCAGCACCGGGTGACTAGGTTGAGCCATGATTAACCTGCAATGA')
     assert len(a) == 20
 
 
-def test_mh_len():
-    a = MinHash(20, 10)
+def test_mh_len(track_abundance):
+    a = MinHash(20, 10, track_abundance=track_abundance)
     for i in range(0, 40, 2):
         a.add_hash(i)
 
     assert a.get_mins() == list(range(0, 40, 2))
 
 
-def test_mh_unsigned_long_long():
-    a = MinHash(20, 10)
+def test_mh_unsigned_long_long(track_abundance):
+    a = MinHash(20, 10, track_abundance=track_abundance)
     a.add_hash(9227159859419181011)        # too big for a C long int.
     assert 9227159859419181011 in a.get_mins()
 
 
-def test_mh_count_common():
-    a = MinHash(20, 10)
+def test_mh_count_common(track_abundance):
+    a = MinHash(20, 10, track_abundance=track_abundance)
     for i in range(0, 40, 2):
         a.add_hash(i)
 
-    b = MinHash(20, 10)
+    b = MinHash(20, 10, track_abundance=track_abundance)
     for i in range(0, 80, 4):
         b.add_hash(i)
 
@@ -182,9 +181,9 @@ def test_mh_count_common():
     assert b.count_common(a) == 10
 
 
-def test_mh_count_common_diff_protein():
-    a = MinHash(20, 5, False)
-    b = MinHash(20, 5, True)
+def test_mh_count_common_diff_protein(track_abundance):
+    a = MinHash(20, 5, False, track_abundance=track_abundance)
+    b = MinHash(20, 5, True, track_abundance=track_abundance)
 
     try:
         a.count_common(b)
@@ -193,9 +192,9 @@ def test_mh_count_common_diff_protein():
         pass
 
 
-def test_mh_count_common_diff_ksize():
-    a = MinHash(20, 5)
-    b = MinHash(20, 6)
+def test_mh_count_common_diff_ksize(track_abundance):
+    a = MinHash(20, 5, track_abundance=track_abundance)
+    b = MinHash(20, 6, track_abundance=track_abundance)
 
     try:
         a.count_common(b)
@@ -204,12 +203,12 @@ def test_mh_count_common_diff_ksize():
         pass
 
 
-def test_mh_asymmetric():
-    a = MinHash(20, 10)
+def test_mh_asymmetric(track_abundance):
+    a = MinHash(20, 10, track_abundance=track_abundance)
     for i in range(0, 40, 2):
         a.add_hash(i)
 
-    b = MinHash(10, 10)                   # different size: 10
+    b = MinHash(10, 10, track_abundance=track_abundance)                   # different size: 10
     for i in range(0, 80, 4):
         b.add_hash(i)
 
@@ -220,13 +219,13 @@ def test_mh_asymmetric():
     assert b.compare(a) == 1.0
 
 
-def test_mh_merge():
+def test_mh_merge(track_abundance):
     # test merging two identically configured minhashes
-    a = MinHash(20, 10)
+    a = MinHash(20, 10, track_abundance=track_abundance)
     for i in range(0, 40, 2):
         a.add_hash(i)
 
-    b = MinHash(20, 10)
+    b = MinHash(20, 10, track_abundance=track_abundance)
     for i in range(0, 80, 4):
         b.add_hash(i)
 
@@ -241,13 +240,13 @@ def test_mh_merge():
 
 
 
-def test_mh_asymmetric_merge():
+def test_mh_asymmetric_merge(track_abundance):
     # test merging two asymmetric (different size) MHs
-    a = MinHash(20, 10)
+    a = MinHash(20, 10, track_abundance=track_abundance)
     for i in range(0, 40, 2):
         a.add_hash(i)
 
-    b = MinHash(10, 10)                   # different size: 10
+    b = MinHash(10, 10, track_abundance=track_abundance)                   # different size: 10
     for i in range(0, 80, 4):
         b.add_hash(i)
 
@@ -263,13 +262,13 @@ def test_mh_asymmetric_merge():
     assert c.compare(b) == 0.5
 
 
-def test_mh_inplace_concat_asymmetric():
+def test_mh_inplace_concat_asymmetric(track_abundance):
     # test merging two asymmetric (different size) MHs
-    a = MinHash(20, 10)
+    a = MinHash(20, 10, track_abundance=track_abundance)
     for i in range(0, 40, 2):
         a.add_hash(i)
 
-    b = MinHash(10, 10)                   # different size: 10
+    b = MinHash(10, 10, track_abundance=track_abundance)                   # different size: 10
     for i in range(0, 80, 4):
         b.add_hash(i)
 
@@ -288,13 +287,13 @@ def test_mh_inplace_concat_asymmetric():
     assert c.compare(b) == 0.5
 
 
-def test_mh_inplace_concat():
+def test_mh_inplace_concat(track_abundance):
     # test merging two identically configured minhashes
-    a = MinHash(20, 10)
+    a = MinHash(20, 10, track_abundance=track_abundance)
     for i in range(0, 40, 2):
         a.add_hash(i)
 
-    b = MinHash(20, 10)
+    b = MinHash(20, 10, track_abundance=track_abundance)
     for i in range(0, 80, 4):
         b.add_hash(i)
 
@@ -308,9 +307,9 @@ def test_mh_inplace_concat():
     assert c.compare(d) == 1.0
     assert d.compare(c) == 1.0
 
-def test_mh_merge_diff_protein():
-    a = MinHash(20, 5, False)
-    b = MinHash(20, 5, True)
+def test_mh_merge_diff_protein(track_abundance):
+    a = MinHash(20, 5, False, track_abundance=track_abundance)
+    b = MinHash(20, 5, True, track_abundance=track_abundance)
 
     try:
         a.merge(b)
@@ -319,9 +318,9 @@ def test_mh_merge_diff_protein():
         pass
 
 
-def test_mh_merge_diff_ksize():
-    a = MinHash(20, 5)
-    b = MinHash(20, 6)
+def test_mh_merge_diff_ksize(track_abundance):
+    a = MinHash(20, 5, track_abundance=track_abundance)
+    b = MinHash(20, 6, track_abundance=track_abundance)
 
     try:
         a.merge(b)
@@ -330,9 +329,9 @@ def test_mh_merge_diff_ksize():
         pass
 
 
-def test_mh_compare_diff_protein():
-    a = MinHash(20, 5, False)
-    b = MinHash(20, 5, True)
+def test_mh_compare_diff_protein(track_abundance):
+    a = MinHash(20, 5, False, track_abundance=track_abundance)
+    b = MinHash(20, 5, True, track_abundance=track_abundance)
 
     try:
         a.compare(b)
@@ -341,9 +340,9 @@ def test_mh_compare_diff_protein():
         pass
 
 
-def test_mh_compare_diff_ksize():
-    a = MinHash(20, 5)
-    b = MinHash(20, 6)
+def test_mh_compare_diff_ksize(track_abundance):
+    a = MinHash(20, 5, track_abundance=track_abundance)
+    b = MinHash(20, 6, track_abundance=track_abundance)
 
     try:
         a.compare(b)
@@ -352,9 +351,9 @@ def test_mh_compare_diff_ksize():
         pass
 
 
-def test_mh_concat_diff_protein():
-    a = MinHash(20, 5, False)
-    b = MinHash(20, 5, True)
+def test_mh_concat_diff_protein(track_abundance):
+    a = MinHash(20, 5, False, track_abundance=track_abundance)
+    b = MinHash(20, 5, True, track_abundance=track_abundance)
 
     try:
         a += b
@@ -363,9 +362,9 @@ def test_mh_concat_diff_protein():
         pass
 
 
-def test_mh_concat_diff_ksize():
-    a = MinHash(20, 5)
-    b = MinHash(20, 6)
+def test_mh_concat_diff_ksize(track_abundance):
+    a = MinHash(20, 5, track_abundance=track_abundance)
+    b = MinHash(20, 6, track_abundance=track_abundance)
 
     try:
         a += b
@@ -374,8 +373,8 @@ def test_mh_concat_diff_ksize():
         pass
 
 
-def test_short_sequence():
-    a = MinHash(20, 5)
+def test_short_sequence(track_abundance):
+    a = MinHash(20, 5, track_abundance=track_abundance)
     a.add_sequence('GGGG')
     # adding a short sequence should fail silently
     assert len(a.get_mins()) == 0
@@ -393,28 +392,28 @@ def test_murmur():
 
 
 def test_abundance_simple():
-    a = MinHash(20, 5, False, True)
+    a = MinHash(20, 5, False, track_abundance=True)
 
     a.add_sequence('AAAAA')
     assert a.get_mins() == [2110480117637990133]
-    assert a.get_mins(True) == {2110480117637990133: 1}
+    assert a.get_mins(with_abundance=True) == {2110480117637990133: 1}
 
     a.add_sequence('AAAAA')
     assert a.get_mins() == [2110480117637990133]
-    assert a.get_mins(True) == {2110480117637990133: 2}
+    assert a.get_mins(with_abundance=True) == {2110480117637990133: 2}
 
 
 def test_abundance_simple_2():
-    a = MinHash(20, 5, False, True)
-    b = MinHash(20, 5, False, True)
+    a = MinHash(20, 5, False, track_abundance=True)
+    b = MinHash(20, 5, False, track_abundance=True)
 
     a.add_sequence('AAAAA')
     assert a.get_mins() == [2110480117637990133]
-    assert a.get_mins(True) == {2110480117637990133: 1}
+    assert a.get_mins(with_abundance=True) == {2110480117637990133: 1}
 
     a.add_sequence('AAAAA')
     assert a.get_mins() == [2110480117637990133]
-    assert a.get_mins(True) == {2110480117637990133: 2}
+    assert a.get_mins(with_abundance=True) == {2110480117637990133: 2}
 
     b.add_sequence('AAAAA')
     assert a.count_common(b) == 1
